multi-bots/requirements.txt
python-telegram-bot==20.8
aiogram==2.*
fastapi==0.111.0
uvicorn==0.30.0
pandas==2.2.2
requests
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
pytz
openpyxl
xlrd
numpy
matplotlib
aiofiles==23.2.1
aiohttp
